Stellar flowtype merged to one file
